+++ packages/dso-toolkit/components/Componenten/dropdown-button/dropdown-button.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  notes: 'Toont een dropdown met daarin een gevulde uitklapper. Gebaseerd op de Bootstrap "Dropdowns" component',
+  notes: 'Toont een dropdown met daarin een gevulde uitklapper. Gebaseerd op de Bootstrap "Dropdowns" component. Class `text-right` op de parent van of op `.dropdown` zelf zorgt voor een rechter uitlijning.',
   status: 'ready',
   collated: true,
   collator: function (markup, item) {
@@ -263,6 +263,150 @@ module.exports = {
         id: 'dropdown_primary_button_open_checkable',
         modifier: 'primary',
         open: true,
+        modifiers: 'dso-checkable',
+        groups: [
+          {
+            label: 'Versies',
+            items: [
+              {
+                label: '10.6.0',
+                checked: true
+              },
+              {
+                label: '10.5.0'
+              },
+              {
+                label: '10.4.0'
+              }
+            ]
+          },
+          {
+            items: [
+              {
+                label: 'master'
+              }
+            ]
+          },
+          {
+            label: 'Branch releases',
+            items: [
+              {
+                label: '#500-Margins-Testbuilds'
+              },
+              {
+                label: '#611-Pager-component-uitbreiden'
+              },
+              {
+                label: '#663-Dropdown-button-toegankelijk-maken'
+              }
+            ]
+          },
+        ]
+      }
+    },
+    {
+      name: 'dropdown-link-button-align-right',
+      context: {
+        id: 'dropdown_link_button_align_right',
+        modifier: 'link',
+        open: true,
+        align: 'right',
+        modifiers: 'dso-checkable',
+        groups: [
+          {
+            label: 'Versies',
+            items: [
+              {
+                label: '10.6.0',
+                checked: true
+              },
+              {
+                label: '10.5.0'
+              },
+              {
+                label: '10.4.0'
+              }
+            ]
+          },
+          {
+            items: [
+              {
+                label: 'master'
+              }
+            ]
+          },
+          {
+            label: 'Branch releases',
+            items: [
+              {
+                label: '#500-Margins-Testbuilds'
+              },
+              {
+                label: '#611-Pager-component-uitbreiden'
+              },
+              {
+                label: '#663-Dropdown-button-toegankelijk-maken'
+              }
+            ]
+          },
+        ]
+      }
+    },
+    {
+      name: 'dropdown-default-button-align-right',
+      context: {
+        id: 'dropdown_default_button_align_right',
+        modifier: 'default',
+        open: true,
+        align: 'right',
+        modifiers: 'dso-checkable',
+        groups: [
+          {
+            label: 'Versies',
+            items: [
+              {
+                label: '10.6.0',
+                checked: true
+              },
+              {
+                label: '10.5.0'
+              },
+              {
+                label: '10.4.0'
+              }
+            ]
+          },
+          {
+            items: [
+              {
+                label: 'master'
+              }
+            ]
+          },
+          {
+            label: 'Branch releases',
+            items: [
+              {
+                label: '#500-Margins-Testbuilds'
+              },
+              {
+                label: '#611-Pager-component-uitbreiden'
+              },
+              {
+                label: '#663-Dropdown-button-toegankelijk-maken'
+              }
+            ]
+          },
+        ]
+      }
+    },
+    {
+      name: 'dropdown-primary-button-align-right',
+      context: {
+        id: 'dropdown_primary_button_align_right',
+        modifier: 'primary',
+        open: true,
+        align: 'right',
         modifiers: 'dso-checkable',
         groups: [
           {