+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## NEXT
+### Added
+* **dso-toolkit** Dropdown menu met uitlijning links en rechts toevoegen ([#961](https://github.com/dso-toolkit/dso-toolkit/issues/#961))
 
 ### Fixed
 * **dso-toolkit:** Accordion anchor styling aanpassingen ([#1008](https://github.com/dso-toolkit/dso-toolkit/issues/1008))